+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to the Pixel Banner plugin will be documented in this file.
 
 ## 🎉 What's New
+### v3.4.0 - 2025-04-19
+#### ✨ Added
+- New `Max Width` setting for banners to control the maximum width of the banner
+- New `Banner Alignment` control to adjust placement when a banner doesn't fill the width of the note
+
 ### v3.3.4 - 2025-04-18
 #### 📦 Updated
 - Fixed background color on server connection retry button